# docs/swagger.yaml
definitions:
  dto.ApplyPermissionBatchDTO:
    properties:
      codes:
        items:
          type: string
        minItems: 1
        type: array
      email:
        type: string
    required:
    - codes
    - email
    type: object
...
  dto.DeletePermissionPayload:
    properties:
      codes:
        items:
          type: string
        type: array
    required:
    - codes
    type: object
...
  dto.ReadPermissionResponse:
    properties:
      code:
        type: string
      description:
        type: string
    type: object
  dto.ReadPermissionsResponse:
    properties:
      permissions:
        items:
          $ref: '#/definitions/dto.ReadPermissionResponse'
        type: array
    type: object
...
  rest_err.Causes:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  rest_err.RestErr:
    properties:
      causes:
        items:
          $ref: '#/definitions/rest_err.Causes'
        type: array
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: Generic administrative API for a Basic Crud
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Basic CRUD API
  version: "1.0"
paths:
...
  /permission/v1/:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of permissions names
      parameters:
      - description: Page number (min 1)
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadPermissionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Read permissions
      tags:
      - Permission
  /permission/v1/apply:
    post:
      consumes:
      - application/json
      description: Apply a batch of permissions to a user by email
      parameters:
      - description: Email and permission codes
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ApplyPermissionBatchDTO'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Apply permissions to user
      tags:
      - Permission
  /permission/v1/read:
    get:
      consumes:
      - application/json
      description: Read permissions by full code name
      parameters:
      - description: Read code (min 4 characters)
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadPermissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Read permission
      tags:
      - Permission
  /permission/v1/search:
    get:
      consumes:
      - application/json
      description: Search permissions by partial or full code name
      parameters:
      - description: Search query (min 4 characters)
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadPermissionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Search permissions
      tags:
      - Permission
  /permission/v1/user/{email}:
    delete:
      consumes:
      - application/json
      description: Remove a batch of permissions from a user by email
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      - description: Batch of permission codes to remove
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.DeletePermissionPayload'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Remove permissions from user
      tags:
      - Permission
    get:
      consumes:
      - application/json
      description: Read all permissions assigned to a user by email
      parameters:
      - in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReadPermissionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      security:
      - BearerAuth: []
      summary: Read user permissions
      tags:
      - Permission
...
securityDefinitions:
  BearerAuth:
    description: 'Enter the token in the format: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"