# internal/app/admin/permission/controller/controller_impl.go
package controller

import (
	"basic-crud-go/internal/app/admin/permission/binding"
	"basic-crud-go/internal/app/admin/permission/dto"
	"basic-crud-go/internal/app/admin/permission/service"
	"basic-crud-go/internal/configuration/rest_err"
	"fmt"
	"github.com/gin-gonic/gin"
	"net/http"
	"strings"
)

const module string = "Admin-Permission-Controller"

type permissionControllerImpl struct {
	service service.PermissionService
}

func NewPermissionController(s service.PermissionService) PermissionController {
	return &permissionControllerImpl{
		service: s,
	}
}

// ReadAll godoc
// @Summary      Read permissions
// @Description  Retrieve a paginated list of permissions names
// @Tags         Permission
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        page   query     int     false   "Page number (min 1)"
// @Param        limit  query     int     false   "Items per page (default: 10)"
// @Success      200   {object}  dto.ReadPermissionsResponse
// @Failure      400   {object}  rest_err.RestErr
// @Failure      500   {object}  rest_err.RestErr
// @Router       /permission/v1/ [get]
func (c *permissionControllerImpl) ReadAll(ctx *gin.Context) {
	req := binding.ValidatePermissionsDTO(ctx)

	perms, err := c.service.ReadAllPermissions(ctx, req.Page, req.Limit)
	if err != nil {
		restErr := rest_err.NewInternalServerError("error fetching permissions", nil)
		ctx.JSON(restErr.Code, restErr)
		return
	}

	var responses []dto.ReadPermissionResponse
	for _, perm := range perms {
		responses = append(responses, dto.ReadPermissionResponse{
			Code:        perm.Code,
			Description: perm.Description,
		})
	}

	ctx.JSON(http.StatusOK, dto.ReadPermissionsResponse{
		Permissions: responses,
	})
}

// Read godoc
// @Summary      Read permission
// @Description  Read permissions by full code name
// @Tags         Permission
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        code   query     string     true   "Read code (min 4 characters)"
// @Success      200   {object}  dto.ReadPermissionResponse
// @Failure      400   {object}  rest_err.RestErr
// @Failure      500   {object}  rest_err.RestErr
// @Router       /permission/v1/read [get]
func (c *permissionControllerImpl) Read(ctx *gin.Context) {
	req, err := binding.ValidateReadPermissioDTO(ctx)
	if err != nil {
		restErr := rest_err.NewBadRequestValidationError("invalid query", []rest_err.Causes{
			rest_err.NewCause("validation", err.Error()),
		})
		ctx.JSON(restErr.Code, restErr)
		return
	}

	perms, err := c.service.ReadByCode(ctx, req.Code)
	if err != nil {
		if strings.Contains(err.Error(), "not found") {
			restErr := rest_err.NewNotFoundError("code not found")
			ctx.JSON(restErr.Code, restErr)
			return
		}
		restErr := rest_err.NewInternalServerError("error fetching permission", nil)
		ctx.JSON(restErr.Code, restErr)
		return
	}

	ctx.JSON(http.StatusOK, dto.ReadPermissionResponse{
		Code:        perms.Code,
		Description: perms.Description,
	})
}

// Search godoc
// @Summary      Search permissions
// @Description  Search permissions by partial or full code name
// @Tags         Permission
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        query   query     string     true   "Search query (min 4 characters)"
// @Success      200   {object}  dto.ReadPermissionsResponse
// @Failure      400   {object}  rest_err.RestErr
// @Failure      500   {object}  rest_err.RestErr
// @Router       /permission/v1/search [get]
func (c *permissionControllerImpl) Search(ctx *gin.Context) {
	req, err := binding.ValidateSearchPermissionDTO(ctx)
	if err != nil {
		restErr := rest_err.NewBadRequestValidationError("invalid query", []rest_err.Causes{
			rest_err.NewCause("validation", err.Error()),
		})
		ctx.JSON(restErr.Code, restErr)
		return
	}

	perms, err := c.service.Search(ctx, req.Query)

	var responses []dto.ReadPermissionResponse
	for _, perm := range perms {
		responses = append(responses, dto.ReadPermissionResponse{
			Code:        perm.Code,
			Description: perm.Description,
		})
	}

	ctx.JSON(http.StatusOK, dto.ReadPermissionsResponse{
		Permissions: responses,
	})
}

// Apply godoc
// @Summary      Apply permissions to user
// @Description  Apply a batch of permissions to a user by email
// @Tags         Permission
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        payload  body      dto.ApplyPermissionBatchDTO  true  "Email and permission codes"
// @Success      204
// @Failure      400   {object}  rest_err.RestErr
// @Failure      500   {object}  rest_err.RestErr
// @Router       /permission/v1/apply [post]
func (c *permissionControllerImpl) Apply(ctx *gin.Context) {
	req, err := binding.ValidateApplyPermissionBatchDTO(ctx)
	if err != nil {
		restErr := rest_err.NewBadRequestValidationError("invalid input", []rest_err.Causes{
			rest_err.NewCause("validation", err.Error()),
		})
		ctx.JSON(restErr.Code, restErr)
		return
	}

	err = c.service.ApplyPermissionUserBatch(ctx, req.Email, req.Codes)
	if err != nil {
		if strings.Contains(err.Error(), "not found") {
			restErr := rest_err.NewNotFoundError(err.Error())
			ctx.JSON(restErr.Code, restErr)
			return
		}
		restErr := rest_err.NewInternalServerError("failed to apply permissions", nil)
		ctx.JSON(restErr.Code, restErr)
		return
	}

	ctx.Status(http.StatusNoContent)
}

// ReadUserPermission godoc
// @Summary      Read user permissions
// @Description  Read all permissions assigned to a user by email
// @Tags         Permission
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        email  path     dto.ReadUserPermissionsDTO  true  "Email of the user"
// @Success      200   {object}  dto.ReadPermissionsResponse
// @Failure      400   {object}  rest_err.RestErr
// @Failure      404   {object}  rest_err.RestErr
// @Failure      500   {object}  rest_err.RestErr
// @Router       /permission/v1/user/{email} [get]
func (c *permissionControllerImpl) ReadUserPermission(ctx *gin.Context) {
	req, err := binding.ValidateReadUserPermissionsDTO(ctx)
	if err != nil {
		restErr := rest_err.NewBadRequestValidationError("invalid request body", []rest_err.Causes{
			rest_err.NewCause("validation", err.Error()),
		})
		ctx.JSON(restErr.Code, restErr)
		return
	}

	permissions, err := c.service.ReadPermissionsUser(ctx, req.Email)
	if err != nil {
		if strings.Contains(err.Error(), "nao encontrada") {
			restErr := rest_err.NewNotFoundError(fmt.Sprintf("user %s not found", req.Email))
			ctx.JSON(restErr.Code, restErr)
			return
		}
		restErr := rest_err.NewInternalServerError("error reading user permissions", nil)
		ctx.JSON(restErr.Code, restErr)
		return
	}

	var responses []dto.ReadPermissionResponse
	for _, perm := range permissions {
		responses = append(responses, dto.ReadPermissionResponse{
			Code:        perm.Code,
			Description: perm.Description,
		})
	}

	ctx.JSON(http.StatusOK, dto.ReadPermissionsResponse{
		Permissions: responses,
	})
}

// RemoveBatch godoc
// @Summary      Remove permissions from user
// @Description  Remove a batch of permissions from a user by email
// @Tags         Permission
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Param        email   path      string                       true  "User email"
// @Param        payload body      dto.DeletePermissionPayload  true  "Batch of permission codes to remove"
// @Success      204
// @Failure      400   {object}  rest_err.RestErr
// @Failure      404   {object}  rest_err.RestErr
// @Failure      500   {object}  rest_err.RestErr
// @Router       /permission/v1/user/{email} [delete]
func (c *permissionControllerImpl) RemoveBatch(ctx *gin.Context) {
	email, codes, err := binding.ValidateDeletePermissionBatchDTO(ctx)
	if err != nil {
		restErr := rest_err.NewBadRequestValidationError("invalid request body", []rest_err.Causes{
			rest_err.NewCause("validation", err.Error()),
		})
		ctx.JSON(restErr.Code, restErr)
		return
	}

	err = c.service.RemovePermissionsBatch(ctx, email, codes)
	if err == nil {
		ctx.Status(http.StatusNoContent)
		return
	}

	if strings.Contains(err.Error(), "not found") || strings.Contains(err.Error(), "nao encontrada") {
		restErr := rest_err.NewNotFoundError(fmt.Sprintf("user %s not found", email))
		ctx.JSON(restErr.Code, restErr)
		return
	}

	restErr := rest_err.NewInternalServerError("error removing permissions", nil)
	ctx.JSON(restErr.Code, restErr)
}